asdf

+++ isola.c
@@ -72,11 +72,11 @@ void isola_get_window(void){
 	if (isola_info_window.width > isola_info_window.height) {
 		isola_info_window.yRatio = 1.;
 		isola_info_window.xRatio = (float)isola_info_window.width
-				/isola_info_window.height;
+				/(float)isola_info_window.height;
 	}else {
 		isola_info_window.xRatio = 1.;
 		isola_info_window.yRatio = (float)isola_info_window.height 
-				/isola_info_window.width;
+				/(float)isola_info_window.width;
 	}
 	isola_info_window.pixelWidth = (double)2./isola_info_window.width;
 	isola_info_window.pixelHeight = (double)2./isola_info_window.height;
